Fix broken --log_to_stderr=false option
Port over missing checks from tornado logging code to gateone logging
code.

It's kind of unfortunate that gateone duplicate tornado logging logic
while still dependending on it. It means this fix is fragile.

Other logging features, like rotation based on date are also broken
because of the duplication and are not addressed by this fix.